+++ js/instantsearch.js/es/lib/utils/unescape.d.ts
@@ -1,9 +1,0 @@
-/**
- * This implementation is taken from Lodash implementation.
- * See: https://github.com/lodash/lodash/blob/4.17.11-npm/unescape.js
- */
-/**
- * Converts the HTML entities "&", "<", ">", '"', and "'" in `string` to their
- * characters.
- */
-export default function unescape(value: string): string;